+++ CHANGELOG.md
@@ -9,8 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as
 straightforward as possible.
+
 ### Added
+- ChangeTracker. Log long expiration keys
 - Progress bar for dump
+
 ### Changed
 
 ### Fixed